requirements.txt
matplotlib==3.1.1
pyserial==3.5
PyYAML==6.0
simple_pid==1.0.1
